.dagger/pyproject.toml
[project]
name = "localstack-dagger-module"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["dagger-io"]

[tool.uv.sources]
dagger-io = { path = "sdk", editable = true }

[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"
